Corrected parameters of timer_set_mode() to u32.

+++ include/libopencm3/stm32/timer.h
@@ -857,8 +857,8 @@ void timer_enable_irq(u32 timer_peripheral, u32 irq);
 void timer_disable_irq(u32 timer_peripheral, u32 irq);
 bool timer_get_flag(u32 timer_peripheral, u32 flag);
 void timer_clear_flag(u32 timer_peripheral, u32 flag);
-void timer_set_mode(u32 timer_peripheral, u8 clock_div,
-		    u8 alignment, u8 direction);
+void timer_set_mode(u32 timer_peripheral, u32 clock_div,
+		    u32 alignment, u32 direction);
 void timer_set_clock_division(u32 timer_peripheral, u32 clock_div);
 void timer_enable_preload(u32 timer_peripheral);
 void timer_disable_preload(u32 timer_peripheral);
